+++ modules/util/stixhelpers.py
@@ -5,6 +5,7 @@ import stix2
 import urllib3
 from modules import site_config
 from . import buildhelpers
+import time
 
 def get_mitigation_list(src):
     """Reads the STIX and returns a list of all mitigations in the STIX"""
@@ -133,73 +134,131 @@ def get_technique_id_domain_map(ms):
     
     return tech_list
 
+def add_replace_or_ignore(stix_objs, attack_id_objs, obj_in_question):
+    """ Add if object does not already exist
+        Replace object if exist depending on deprecation status or modified date
+        Ignore if object already exists but object in question is outdated
+
+        Deconflicts objects by ATT&CK and STIX IDs
+    """
+
+    def has_STIX_ATTACK_ID_conflict(attack_id):
+        # Check if STIX ID has been seen before, if it has, return ATT&CK ID of conflict ATT&CK if ATT&CK IDs are different
+        conflict = stix_objs.get(obj_in_question.id)
+        if conflict:
+            conflict_attack_id = buildhelpers.get_attack_id(conflict)
+            if conflict_attack_id != attack_id and attack_id_objs.get(conflict_attack_id):
+                return conflict_attack_id
+        
+        return None
+    
+    def replace_object(attack_id, conflict_attack_id):
+        # Replaces object on ATT&CK and STIX maps
+        # Verify for STIX to ATT&CK conflict
+        if conflict_attack_id:
+            attack_id_objs[attack_id] = obj_in_question
+            # Remove outdated ATT&CK ID from map
+            attack_id_objs.pop(conflict_attack_id)
+        else:
+            attack_id_objs[attack_id] = obj_in_question
+
+        stix_objs[obj_in_question.id] = obj_in_question
+
+    # Get ATT&CK ID
+    attack_id = buildhelpers.get_attack_id(obj_in_question)
+
+    if not attack_id:
+        # Ignore if ATT&CK ID does not exist
+        return
+
+    # Get ATT&CK ID if there is possible conflict with STIX ID and ATT&CK ID
+    conflict_attack_id = has_STIX_ATTACK_ID_conflict(attack_id)
+
+    # Check if object in conflict exists
+    # STIX ID
+    stix_id_obj_in_conflict = stix_objs.get(obj_in_question.id)
+
+    # Get ATT&CK ID conflicts
+    if conflict_attack_id:
+        attack_id_obj_in_conflict = attack_id_objs.get(conflict_attack_id)
+    else:
+        attack_id_obj_in_conflict = attack_id_objs.get(attack_id)
+
+    # ------------------------------------------------------------------------
+    # Add: Object does not exist
+
+    if not stix_id_obj_in_conflict:
+        # Add if object does not exist in STIX ID map
+        stix_objs[obj_in_question.id] = obj_in_question
+
+    if not attack_id_obj_in_conflict:
+        # Add if object does not exist in ATT&CK ID map
+        attack_id_objs[attack_id] = obj_in_question
+
+    # ------------------------------------------------------------------------
+    # Replace: Object already exists
+
+    # Ignore if object in question is deprecated and object in conflict is not
+    elif not attack_id_obj_in_conflict.get('x_mitre_deprecated') and obj_in_question.get('x_mitre_deprecated'):
+        return
+
+    # If object in conflict is deprecated and recent object is not, select recent
+    elif attack_id_obj_in_conflict.get('x_mitre_deprecated') and not obj_in_question.get('x_mitre_deprecated'):
+        # Replace object in conflict with object in question
+        replace_object(attack_id, conflict_attack_id)
+
+    # Replace if modified date is more recent
+    else:
+        conflict_modified = attack_id_obj_in_conflict.get('modified')
+        in_question_modified = obj_in_question.get('modified')
+
+        if in_question_modified > conflict_modified:
+            # Replace object in conflict with object in question
+            replace_object(attack_id, conflict_attack_id)
+    
 def grab_resources(ms):
     """Returns a dict that contains lists for the software, group,
        technique and mitigation objects.
     """
 
+    def get_domain_resources(types):
+        # Returns sorted list by name of domain resources by given type list
+        # Builds list from unique ATT&CK IDs
+
+        def grab_filtered_list_by_type(domain, stix_type):
+            return ms[domain['name']].query([
+                stix2.Filter('type', '=', stix_type),
+                stix2.Filter('revoked', '=', False)
+            ])
+
+        # Track objects by STIX ID
+        stix_objs = {}
+        # Track objects by ATT&CK ID
+        attack_id_objs = {}
+        for domain in site_config.domains:
+            if domain['deprecated']: continue
+
+            for stix_type in types:
+                curr_list = grab_filtered_list_by_type(domain, stix_type)
+
+                for val in curr_list:
+                    add_replace_or_ignore(stix_objs, attack_id_objs, val)
+
+        # Convert into list of values
+        resource_list = [attack_id_objs[key] for key in attack_id_objs]
+        return sorted(resource_list, key=lambda k: k['name'].lower())
+
     #Generates the list of techniques
-    tech_list = []
-    for domain in site_config.domains:
-        if domain['deprecated']: continue
-        curr_list = ms[domain['name']].query([
-            stix2.Filter('type', '=', 'attack-pattern'),
-            stix2.Filter('revoked', '=', False)
-        ])
-        for val in curr_list:
-            if next((x for x in tech_list if x.id == val.id), None) is None:
-                tech_list.append(val)   
-    tech_list = sorted(tech_list, key=lambda k: k['name'].lower())
+    tech_list = get_domain_resources(['attack-pattern'])
 
     #Generates list of software
-    software_list = []
-    for domain in site_config.domains:
-        if domain['deprecated']: continue
-        curr_list = ms[domain['name']].query([
-            stix2.Filter('type', '=', 'malware'),
-            stix2.Filter('revoked', '=', False)
-        ])
-        # Remove duplicates from different domains
-        for val in curr_list:
-            if next((x for x in software_list if x.id == val.id), None) is None:
-                software_list.append(val)
-        curr_list = ms[domain['name']].query([
-            stix2.Filter('type', '=', 'tool'),
-            stix2.Filter('revoked', '=', False)
-        ])
-        # Remove duplicates from different domains
-        for val in curr_list:
-            if next((x for x in software_list if x.id == val.id), None) is None:
-                software_list.append(val)   
-    software_list = sorted(software_list, key=lambda k: k["name"].lower() )
+    software_list = get_domain_resources(['malware', 'tool'])
 
     #Generates list of groups
-    group_list = []
-    for domain in site_config.domains:
-        if domain['deprecated']: continue
-        curr_list = ms[domain['name']].query([
-            stix2.Filter('type', '=', 'intrusion-set'),
-            stix2.Filter('revoked', '=', False)
-        ])
-        # Remove duplicates from different domains
-        for val in curr_list:
-            if next((x for x in group_list if x.id == val.id), None) is None:
-                group_list.append(val)
-    group_list = sorted(group_list, key=lambda k: k["name"].lower())
+    group_list = get_domain_resources(['intrusion-set'])
 
     #Generates a list of CoA
-    coa_list = []
-    for domain in site_config.domains:
-        if domain['deprecated']: continue
-        curr_list = ms[domain['name']].query([
-            stix2.Filter("type", "=", "course-of-action"),
-            stix2.Filter('revoked', '=', False)
-        ])
-        #remove duplicates from different domains
-        for val in curr_list:
-            if next((x for x in coa_list if x.id == val.id), None) is None:
-                coa_list.append(val)
-    coa_list = sorted(coa_list, key=lambda k: k["name"].lower() )
+    coa_list = get_domain_resources(['course-of-action'])
 
     #Generates list of relationships
     rel_list = []